Add pump speed control support (0.0-2.0)
Core Changes:
- physical_process.py: Use float() instead of int() to preserve decimal speed values
- control.py: Updated all applyScadaDecision methods to pass through numeric values
- control.py: Added else clause to all applyHybridDecision methods for speed values
- generic_scada.py: Added handling for numeric speed values in SCADA control

Pump Speed Feature:
- Custom algorithms can now return float values (0.0-2.0) for pump speed
- 0.0 = OFF, 1.0 = 100% speed, 1.5 = 150%, 2.0 = max
- Uses EPANET SETTING property for variable speed pumps
- Fully backward compatible with 'open'/'closed' string values

New Example:
- examples/EdenTown/PumpSpeed_CustomAlgo_Example/
- Demonstrates pump speed cycling from 0.0 to 1.5 by 0.1 each iteration
- Includes README with usage instructions

+++ wacsim/physical_process.py
@@ -409,9 +409,10 @@ class PhysicalPlant:
         for control in self.control_list:
             rows_1 = c.execute('SELECT value FROM plant WHERE name = ?', (control['name'],)).fetchone()
             conn.commit()
-            new_status = int(rows_1[0])
+            # Use float() instead of int() to preserve decimal values for pump speed control
+            new_status = float(rows_1[0])
 
-            control['value'] = float(new_status)
+            control['value'] = new_status
             idx = en.getlinkindex(ph=self.proj, id=control['name'])
             if not math.isclose(control['value'], 1) and not math.isclose(control['value'], 0):
                 en.setlinkvalue(ph=self.proj, index=idx, property=en.SETTING, value=control['value'])
